Fix AWS permissions
As we added the name tag to the scanner instances we had to allow the
Name tag key in the policy.

For Scanner debug we must also allow instances to be created with a
key-pair from our account.

We also need to allow CopySnapshot actions in the policy for snapshots
in other regions that must be moved to the VMClarity region.