Redo database interface to use model as input and output
This commit revisits the database interface such that each function takes
in and returns a VMClarity API model object. How the DB driver decides
to store these objects and process them efficently is the resposiblity
of the DB driver. We start with one DB driver, gorm, which converts the
model objects to GORM structs to save them.