Fix $select on primitive collection
When implementing $orderby the aggregation for a collection was broken
into two peices this meant we needed to convert the column data back to
JSON in the aggregate function if its an object otherwise we end up with
an array of string encoded json objects. This only applies to objects
though, so we need to skip the conversion if building an aggregate of
primitive types.