Move default for maxParallelScanners out of ScanConfigData
A value can not have a default and also be readOnly because the default
gets applied if the value is unset so then you are implictly trying to
write it as the default.

This moves the default to the ScanConfig object which is used for
reading/writing, and out of ScanConfigData which is shared by
ScanConfigRelationship and needs all the fields to be readOnly.