Make array types in API nullable
This change makes array types in the OAS nullable, without it when you
create an API request object during a PATCH like:

{
  "Licenses": null
}

the API returns an error that the value is not nullable. We may also
want to do this for the other fields too.

We may want to revisit this when we revisit all the schemas for
GET/POST/PATCH because often null isn't a valid setting for some of
these arrays and it would be better to have a default of an empty list.