Add DB indexes to increase performace
Adding simple indexes like the ID of each of the objects, or some of the
commonly queried fields increases the performance of the SQL queries
because its not having to scan over every object in the DB for the WHERE
queries.

This commit adds indexes for all the object IDs, and also indexes for
common queries performed by the scan result processor and UI on
findings.