Handle upgrade of cloud formation with different defaults
When you deploy a cloud formation stack, even if you leave the default
parameters the values are saved.

When you upgrade the saved values are used instead of the defaults,
which means that if we change the container version in the default for a
parameter it gets ignored on upgrade.

This commit changes the symantics of the parameters to overrides, only
if provided will they override the value which we set in the CFN
template. This means that:

- if you want to return to the default value you can delete the contents
  of the parameter.

- during upgrades the cloud formation will template in the new values
  unless an override is provided.