Ensure that the reconcile queue tracks processing items
This commit adds some new state and functions to the orchestration queue
so that it keeps track of Dequeued items until that item is marked as
Done by the person who Dequeued it.

This information is used to prevent items from being re-queued while the
system is processing an item because the poller feeding the queue may
pick up an item which is in the process of being updated.

This also makes it safe to have multiple reconciler threads being fed
from a single queue because we have a guarentee that the same item isn't
in the queue twice and therefore we can't end up with mutiple
reconcilers trying to process the same item.