Add EnqueueAfter and RequeueAfter option to queue
This commit adds an option for an item to be enqueued after a specified
duration. During that duration the item can not be enqueued if requested
through Enqueue to ensure that the item is not enqueued before that
duration is completed.

This allows reconcilers to delay reconciliation for an item when it
knows it will be performing an operation which takes longer than the
polling cycle preventing the poller from requeuing that item.

RequeueAfter is a wrapper around Done() followed by EnqueueAfter() done
under the same lock to prevent race condition when an item could be
requeued between Done and the EnqueueAfter call.

Also adds a new error type, RequeueAfterError, which can be returned by
a reconcile function to indicate that an item should be requeued after a
certain duration. This should only be used as an optimisation and the
pollers and reconcilers should be designed to pick up correctly from a
restart of the system.