Separate asset discovery from scanning
This commit removes the concept of scopes and instead replaces it with
periodic discovery of assets. These assets will be created in the
system and a ScanConfig can be configured with an ODATA $filter
expression to choose the assets to scan.

This increases user visibility into the assets available to scan by
VMClarity, it also gives the system an opportunity to estimate how many
items a ScanConfig might scan based on a particular query.

It also opens the door for additional features such as the ability to
scan a specific asset which has never been scanned before.