Ensure we cast to JSON while build JSON objects
This commit ensures that while building the json objects for select and
expand, we cast the extracted data to json to avoid issues where the DB
gets confused between JSON encoded strings and JSON data.

Without this change we sometimes get unexpected escaped quotes in the
returned data as the DB thinks its been given a string, but its actually
a json encoded string.