fix: ScanResults stuck in aborted state
The `vmclarity-cli` running on the `Scanner` instance is responsible
for handling the `aborted` event for the corresponding `ScanResult`.
The `ScanResultWatcher` controller will periodically check whether
the `vmclarity-cli` managed to gracefully shutdown the running scan
and set the` ScanResult `to `done` indicating that thr scan is finished
and the Scanner instance could be removed. However moving
the `ScanResults` state to `done` will not happen in case
the `vmclarity-cli` process is not running on the Scanner instance
making the `ScanResultWatcher` controller waiting indefinitely.

The `ScanResultWatcher` will now monitor the `ScanResults` in `aborted`
state and will move their state to `done` after a predefined
period of time.