Volume mount scanner hostfs into scanner container
Containers can only be aware of device mount points which are within
their containerized filesystem. This means that if a device is mounted
somewhere on the host, and that location on the host is not exposed to
the container, the container can not determine if its mount or not.

This commit resolves this issue by making the whole host filesystem
available for the container to see by mounting it into a sub-directory
called `/hostfs` within the containers filesystem.

Once the container can see the host filesystem, all volumes which are
mounted by the host or by the scanner have a mount point defined in
`lsblk`.