feat: introduce helm chart for kubernetes installation
This commit adds a helm chart for installing the VMClarity control plane on
Kubernetes. Communications between the control plane services is
performed internal to the cluster. Postgresql is used as the database
which allows the API Server, UI and UI Backend services to be scaled
horizontally.

While the control plane runs on Kubernetes the provider is still
responsible for where the scanner instances run, and how they
communicate back to the control plane.